* moved the body of some methods to new file _client.py
